Consolidate image pruning and use standard watchdog tests
