+++ test/requirements.txt
@@ -1,4 +1,6 @@
 marshmallow_sqlalchemy
+docker==7.2.0
+psycopg2-binary
 pytest
 pytest-dependency
 pytest-order
