+++ test/test_dojo_cli.py
@@ -1100,7 +1100,8 @@ def test_watchdog_cron_runs():
 
     crontab = dojo_run("docker", "exec", "watchdog", "crontab", "-l").stdout
     assert re.search(r"^\*/5 \* \* \* \* /usr/local/bin/docker_remove_containers", crontab, re.M), crontab
-    assert re.search(r"^0 9 \* \* \* /usr/local/bin/docker_prune_images", crontab, re.M), crontab
+    assert re.search(r"^17 \* \* \* \* /usr/local/bin/docker_prune_images --apply", crontab, re.M), crontab
+    assert crontab.count("/usr/local/bin/docker_prune_images") == 1, crontab
 
     if uptime < 400:
         pytest.skip("watchdog has not been up for a full cron interval")
@@ -1120,10 +1121,10 @@ def test_watchdog_cron_runs():
         assert 240 <= interval <= 360, f"consecutive reaper runs were {interval}s apart"
 
 
-@pytest.mark.skipif(MULTINODE, reason="the pruner targets the workspace nodes, not the main daemon")
-def test_watchdog_prunes_dangling_images():
+def test_watchdog_prunes_unreferenced_images(admin_session):
     target = WORKER_CONTAINER if MULTINODE else DOJO_CONTAINER
     tag = f"cli-prune-{_rand()}"
+    referenced = f"{tag}-referenced"
     existing = set(dojo_run("docker", "images", "-qf", "dangling=true", container=target).stdout.split())
     created = set()
     try:
@@ -1134,17 +1135,34 @@ def test_watchdog_prunes_dangling_images():
         created = set(dojo_run("docker", "images", "-qf", "dangling=true",
                                container=target).stdout.split()) - existing
         assert created, "failed to create a dangling image"
+        dojo_run("sh", "-c",
+                 f'printf "FROM hello-world\\nLABEL probe=referenced\\n" | docker build -q -t {referenced} -',
+                 container=target)
+        create_dojo_yml(
+            f"id: {tag}\nimage: {referenced}\nmodules:\n  - id: hello\n    challenges:\n      - id: apple\n",
+            session=admin_session,
+        )
 
-        result = dojo_run("docker", "exec", "watchdog", "/usr/local/bin/docker_prune_images", check=False)
+        dry_run = dojo_run("docker", "exec", "watchdog", "/usr/local/bin/docker_prune_images")
+        output = dry_run.stdout + dry_run.stderr
+        assert f"would remove {tag}:latest " in output, output
+        assert f"would remove {referenced}:latest " not in output, output
+        assert dojo_run("docker", "image", "inspect", tag, check=False,
+                        container=target).returncode == 0, "dry run removed an image"
+
+        result = dojo_run("docker", "exec", "watchdog", "/usr/local/bin/docker_prune_images", "--apply", check=False)
         output = result.stdout + result.stderr
         assert result.returncode == 0, output[-2000:]
-        assert "Prune docker images complete" in output, output[-2000:]
+        assert ": complete" in output, output[-2000:]
 
         remaining = set(dojo_run("docker", "images", "-qf", "dangling=true", container=target).stdout.split())
         assert not created & remaining, f"dangling images survived the prune: {created & remaining}"
-        assert dojo_run("docker", "image", "inspect", "pwncollege/challenge-simple", check=False,
-                        container=target).returncode == 0, "the pruner removed a tagged image"
+        assert dojo_run("docker", "image", "inspect", tag, check=False,
+                        container=target).returncode != 0, "unreferenced tagged image survived"
+        assert dojo_run("docker", "image", "inspect", referenced, check=False,
+                        container=target).returncode == 0, "the pruner removed a challenge-referenced image"
     finally:
+        dojo_run("docker", "rmi", "-f", referenced, check=False, container=target)
         dojo_run("docker", "rmi", "-f", tag, check=False, container=target)
         for image in created:
             dojo_run("docker", "rmi", "-f", image, check=False, container=target)
